pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weather-vessel"
version = "0.2.0"
description = "Marine logistics weather intelligence CLI"
requires-python = ">=3.11"
authors = [{name = "Weather Vessel"}]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
]
dynamic = ["readme"]

[tool.setuptools.dynamic]
readme = {file = ["README.md"], content-type = "text/markdown"}

[project.urls]
Homepage = "https://github.com/macho715/WEATHER-VESSEL"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=24.1",
    "flake8>=6.1",
    "isort>=5.12",
    "mypy>=1.6",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["wv"]

[tool.coverage.run]
branch = true
source = ["src/wv"]

[tool.coverage.report]
omit = ["tests/*"]
fail_under = 70.0
